chore: Enable multi-os golanglint-ci (#13599)
(cherry picked from commit 9ede0b0c4995a088ecb3d535ea60356662412618)
