+++ plugins/inputs/ethtool/ethtool_linux.go
@@ -19,7 +19,43 @@ import (
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
 
+const (
+	tagInterface     = "interface"
+	tagNamespace     = "namespace"
+	tagDriverName    = "driver"
+	fieldInterfaceUp = "interface_up"
+)
+
 var downInterfacesBehaviors = []string{"expose", "skip"}
+
+type Ethtool struct {
+	// This is the list of interface names to include
+	InterfaceInclude []string `toml:"interface_include"`
+
+	// This is the list of interface names to ignore
+	InterfaceExclude []string `toml:"interface_exclude"`
+
+	// Behavior regarding metrics for downed interfaces
+	DownInterfaces string `toml:" down_interfaces"`
+
+	// This is the list of namespace names to include
+	NamespaceInclude []string `toml:"namespace_include"`
+
+	// This is the list of namespace names to ignore
+	NamespaceExclude []string `toml:"namespace_exclude"`
+
+	// Normalization on the key names
+	NormalizeKeys []string `toml:"normalize_keys"`
+
+	Log telegraf.Logger `toml:"-"`
+
+	interfaceFilter   filter.Filter
+	namespaceFilter   filter.Filter
+	includeNamespaces bool
+
+	// the ethtool command
+	command Command
+}
 
 type CommandEthtool struct {
 	Log                 telegraf.Logger
